backend/fastapi/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
